+++ spec/celluloid/actor/system_spec.rb
@@ -73,7 +73,7 @@ RSpec.describe Celluloid::Actor::System do
     subject.shutdown
 
     expect { subject.get_thread }
-      .to raise_error(Celluloid::Group::NotActive)
+      .to raise_error(Celluloid::NotActive)
   end
 
   it "warns nicely when no actor system is started" do