+++ spec/celluloid/condition_spec.rb
@@ -68,12 +68,12 @@ RSpec.describe Celluloid::Condition, actor_system: :global do
   it "times out inside normal Threads" do
     condition = Celluloid::Condition.new
     expect { condition.wait(1) }
-      .to raise_error(Celluloid::ConditionError)
+      .to raise_error(Celluloid::ConditionTimedOut)
   end
 
   it "times out inside Tasks" do
-    allow(logger).to receive(:crash).with("Actor crashed!", Celluloid::ConditionError)
+    allow(logger).to receive(:crash).with("Actor crashed!", Celluloid::ConditionTimedOut)
     expect { actor.wait_for_condition(1) }
-      .to raise_error(Celluloid::ConditionError)
+      .to raise_error(Celluloid::ConditionTimedOut)
   end
 end